# Properly set inbound and outbound connection and stream window sizes